Add script to create regional tile files
Unlike runGlobal.sh, which is a command line tool that accepts all
configurables as passed arguments, runRegional.template.sh should be
edited and run on its own for each inventory. This is because it relies
on the bounds and steps of latitude and longitude in the source lat/lon
file which are easy to make a mistake entering. This script is
thus designed to be copied and archived for each new inventory for easy
checking and editing later on as needed.

As with runGlobal.sh, output netcdf files are put in a TileFiles
subdirectory. The filename is consistent with global filename formats
except that DE/DC and PE/PC are replaced with UU.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>